src/backend/requirements.txt
fastapi
uvicorn
python-multipart
psycopg2-binary
pgvector
pydantic
pydantic-settings
numpy
opencv-python-headless
insightface
ultralytics
torch
